Splits static syntax tree analysis from transformations.
PiperOrigin-RevId: 258201908
