+++ docs/api_docs/python/tff/framework/check_intrinsics_whitelisted_for_reduction.md
@@ -11,7 +11,7 @@ Checks whitelist of intrinsics reducible to aggregate or broadcast.
 tff.framework.check_intrinsics_whitelisted_for_reduction(comp)
 ```
 
-<a target="_blank" href=http://github.com/tensorflow/federated/tree/master/tensorflow_federated/python/core/impl/transformations.py>View
+<a target="_blank" href=http://github.com/tensorflow/federated/tree/master/tensorflow_federated/python/core/impl/tree_analysis.py>View
 source</a>
 
 <!-- Placeholder for "Used in" -->
