+++ tensorflow_federated/python/core/impl/transformations.py
@@ -19,7 +19,6 @@ from __future__ import division
 from __future__ import print_function
 
 import itertools
-
 import six
 from six.moves import range
 from six.moves import zip
@@ -35,8 +34,8 @@ from tensorflow_federated.python.core.impl import context_stack_base
 from tensorflow_federated.python.core.impl import federated_computation_utils
 from tensorflow_federated.python.core.impl import intrinsic_bodies
 from tensorflow_federated.python.core.impl import intrinsic_defs
-from tensorflow_federated.python.core.impl import placement_literals
 from tensorflow_federated.python.core.impl import transformation_utils
+from tensorflow_federated.python.core.impl import tree_analysis
 from tensorflow_federated.python.core.impl import type_utils
 
 
@@ -86,7 +85,7 @@ def extract_intrinsics(comp):
   """
   py_typecheck.check_type(comp,
                           computation_building_blocks.ComputationBuildingBlock)
-  check_has_unique_names(comp)
+  tree_analysis.check_has_unique_names(comp)
   name_generator = computation_constructing_utils.unique_name_generator(comp)
   unbound_references = get_map_of_unbound_references(comp)
 
@@ -294,7 +293,7 @@ def inline_block_locals(comp, variable_names=None):
   """
   py_typecheck.check_type(comp,
                           computation_building_blocks.ComputationBuildingBlock)
-  check_has_unique_names(comp)
+  tree_analysis.check_has_unique_names(comp)
   if variable_names is not None:
     py_typecheck.check_type(variable_names, (list, tuple, set))
 
@@ -1274,36 +1273,6 @@ def insert_called_tf_identity_at_leaves(comp):
       comp, _decorate_if_reference_without_graph)
 
 
-def check_has_single_placement(comp, single_placement):
-  """Checks that the AST of `comp` contains only `single_placement`.
-
-  Args:
-    comp: Instance of `computation_building_blocks.ComputationBuildingBlock`.
-    single_placement: Instance of `placement_literals.PlacementLiteral` which
-      should be the only placement present under `comp`.
-
-  Raises:
-    ValueError: If the AST under `comp` contains any
-    `computation_types.FederatedType` other than `single_placement`.
-  """
-  py_typecheck.check_type(comp,
-                          computation_building_blocks.ComputationBuildingBlock)
-  py_typecheck.check_type(single_placement, placement_literals.PlacementLiteral)
-
-  def _check_single_placement(comp):
-    """Checks that the placement in `type_spec` matches `single_placement`."""
-    if (isinstance(comp.type_signature, computation_types.FederatedType) and
-        comp.type_signature.placement != single_placement):
-      raise ValueError(
-          'Comp contains a placement other than {}; '
-          'placement {} on comp {} inside the structure. '.format(
-              single_placement, comp.type_signature.placement,
-              computation_building_blocks.compact_representation(comp)))
-    return comp, False
-
-  transformation_utils.transform_postorder(comp, _check_single_placement)
-
-
 def unwrap_placement(comp):
   """Strips `comp`'s placement, returning a single call to map, apply or value.
 
@@ -1349,7 +1318,7 @@ def unwrap_placement(comp):
   py_typecheck.check_type(comp.type_signature, computation_types.FederatedType)
   single_placement = comp.type_signature.placement
 
-  check_has_single_placement(comp, single_placement)
+  tree_analysis.check_has_single_placement(comp, single_placement)
 
   name_generator = computation_constructing_utils.unique_name_generator(comp)
 
@@ -1583,15 +1552,6 @@ def replace_all_intrinsics_with_bodies(comp, context_stack):
   return comp, transformed
 
 
-def check_has_unique_names(comp):
-  if not transformation_utils.has_unique_names(comp):
-    raise ValueError(
-        'This transform should only be called after we have uniquified all '
-        '`computation_building_blocks.Reference` names, since we may be moving '
-        'computations with unbound references under constructs which bind '
-        'those references.')
-
-
 def get_map_of_unbound_references(comp):
   """Gets a Python `dict` of the unbound references in `comp`.
 
@@ -1643,44 +1603,3 @@ def get_map_of_unbound_references(comp):
 
   transformation_utils.transform_postorder(comp, _update)
   return references
-
-
-def check_intrinsics_whitelisted_for_reduction(comp):
-  """Checks whitelist of intrinsics reducible to aggregate or broadcast.
-
-  Args:
-    comp: Instance of `computation_building_blocks.ComputationBuildingBlock` to
-      check for presence of intrinsics not currently immediately reducible to
-      `FEDERATED_AGGREGATE` or `FEDERATED_BROADCAST`, or local processing.
-
-  Raises:
-    ValueError: If we encounter an intrinsic under `comp` that is not
-    whitelisted as currently reducible.
-  """
-  # TODO(b/135930668): Factor this and other non-transforms (e.g.
-  # `check_has_unique_names` out of this file into a structure specified for
-  # static analysis of ASTs.
-  py_typecheck.check_type(comp,
-                          computation_building_blocks.ComputationBuildingBlock)
-  uri_whitelist = (
-      intrinsic_defs.FEDERATED_AGGREGATE.uri,
-      intrinsic_defs.FEDERATED_APPLY.uri,
-      intrinsic_defs.FEDERATED_BROADCAST.uri,
-      intrinsic_defs.FEDERATED_MAP.uri,
-      intrinsic_defs.FEDERATED_MAP_ALL_EQUAL.uri,
-      intrinsic_defs.FEDERATED_VALUE_AT_CLIENTS.uri,
-      intrinsic_defs.FEDERATED_VALUE_AT_SERVER.uri,
-      intrinsic_defs.FEDERATED_ZIP_AT_SERVER.uri,
-      intrinsic_defs.FEDERATED_ZIP_AT_CLIENTS.uri,
-  )
-
-  def _check_whitelisted(comp):
-    if isinstance(comp, computation_building_blocks.Intrinsic
-                 ) and comp.uri not in uri_whitelist:
-      raise ValueError(
-          'Encountered an Intrinsic not currently reducible to aggregate or '
-          'broadcast, the intrinsic {}'.format(
-              computation_building_blocks.compact_representation(comp)))
-    return comp, False
-
-  transformation_utils.transform_postorder(comp, _check_whitelisted)
